Compare assert_near values in float64 instead of float32
Both ``assert_near`` implementations (``tools/__init__.py`` and
``tools/test_runner.py``) downcast their operands to ``np.float32``
before computing the diff. Float32 only carries ~7 decimal digits, so
any pair above ~16M that happened to round to the same float32 was
treated as equal even with ``absolute_error_margin=0``. A test expecting
25_000_001 passed silently on 25_000_000.

Compare in float64 so dollar-level differences on multi-million-dollar
values are caught.

One existing tax-scale test was relying on float32 rounding making 0.3
look exact. Relax its absolute_error_margin from 0 to 1e-10 and note
the behaviour change in the PR body.

Fixes H6 from the 2026-04 bug hunt.

+++ policyengine_core/tools/__init__.py
@@ -43,9 +43,13 @@ def assert_near(
     if isinstance(target_value, str):
         target_value = eval_expression(target_value)
 
-    target_value = np.array(target_value).astype(np.float32)
+    # Use float64 here so we don't silently lose precision on values
+    # above ~16M (float32 only carries ~7 decimal digits). Under float32,
+    # ``25_000_001`` and ``25_000_000`` round to the same number and a
+    # test expecting one would pass on the other (bug H6).
+    target_value = np.array(target_value).astype(np.float64)
 
-    value = np.array(value).astype(np.float32)
+    value = np.array(value).astype(np.float64)
     diff = abs(target_value - value)
     if absolute_error_margin is not None:
         assert (diff <= absolute_error_margin).all(), (
